+++ CHANGELOG.md
@@ -9,8 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+
+## [1.2.4] 2024-07-12
+
+### Added
+
 * Added `list` to accepted types for `Scene.add`.
 * Added `list[float]` to accepted types for `Camera.position` and `Camera.target`.
+* Added `unit` to `Viewer` and `Config`.
+* Added `bounding_box` and `_update_bounding_box` to `BufferObject`.
 * Added `TextEdit` to handle `name` change.
 * Added `DefaultLayout` to handle gerneral `layout` setting to minimal.
 * Added `ColorDialog` to manage color dialog.
@@ -20,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Fixed `opacity` bug with `BufferObject`.
+* Updated `SceneForm` to avoid completely reload when scene objects not changed.
 * Updated callback to `SceneTree`.
 * Updated `ObjectSetting` and `CameraSetting` to support setting from config.
 * Updated `Slider` to be able change value with `TextEdit`